Promote MySQL Connector/J to regular dependency
Set the MySQL Connector/J JDBC driver to being a regular dependency.
This means it will be included in this project's uberjar whenever the
uberjar is being built (e.g. when the uberjar is built by Leiningen).
This means the uberjar now constitutes a redistribution of the MySQL
Connector/J JDBC driver. Following Oracle's FOSS License Exception
applying to this driver, this is fine as long as this project is
licensed under the EPL 1.0 (or any other license included in that FOSS
exception), and as long as the terms of that FOSS License Exception are
fulfilled. The Foss License Exception applying to this driver is
currently published by Oracle at
http://www.mysql.com/about/legal/licensing/foss-exception/.